Clean project files.
