+++ MazeWPF/NodeState.cs
@@ -1,0 +1,10 @@
+﻿namespace MazeWPF
+{
+    public enum NodeState
+    {
+        Current,
+        Backtracked,
+        Unvisited,
+        Visited
+    }
+}
